Methods list

Sinograms.Sinograms

Sinograms.fbp

Sinograms.fbp!

Sinograms.fbp2

Sinograms.fbp2

Sinograms.fbp2_back

Sinograms.fbp2_back_fan

Sinograms.fbp2_sino_filter

Sinograms.fbp2_sino_weight

Sinograms.fbp2_window

Sinograms.fbp_ramp

Methods usage

Sinograms.Sinograms — Module

Sinograms module

Sinograms.fbp! — Method

fbp!(image, sino ; orbit::Real = 180, kwargs...)

FBP reconstruction from a parallel-beam sinogram of size [nr × nϕ]. Writes result into image matrix.

Input

sino::AbstractMatrix

Options for sino_geom

dr : sinogram radial spacing; default 1

orbit : angular range in degrees; default 180

orbit_start : angular range in degrees; default 180

Options for image_geom

dx, dy, offset_x, offset_y

Options

kwargs : passed to fbp2

Output

image::AbstractMatrix is mutated

Sinograms.fbp — Method

fbp(sino::AbstractMatrix ; kwargs...)

FBP reconstruction from a parallel-beam sinogram of size [nr × nϕ]. Returns an image of size [nx × ny].

Options

nx : default nr

ny : default nx

kwargs : passed to fbp2

Sinograms.fbp2 — Method

plan = fbp2(sg, ig; how=:normal, window=:none)

FBP 2D tomographic image reconstruction for parallel-beam or fan-beam cases, with either flat or arc detector for fan-beam case.

To use this, you first call it with the sinogram and image geometry. The routine returns the initialized "plan". Thereafter, to to perform FBP reconstruction, you call this routine with the plan (perhaps numerous times for the same geometry).

in

sg::SinoGeom

ig::ImageGeom

options

how::Symbol how to reconstruct

:normal default

:mojette use mojette rebinning and Gtomo2_table

window::Symbol e.g. :hann (default: :none)

-T::DataType type of sino elements (default: Float32)

out

plan::FBPplan initialized plan

Sinograms.fbp2 — Method

image, sino_filt = fbp2(plan, sino)

Filtered back-projection (FBP) reconstruction, returning image and filtered sinogram

in

plan::FBPplan

sino::AbstractMatrix{<:Number}

out

image::AbstractMatrix{<:Number} reconstructed image(s)

sino_filt::AbstractMatrix{<:Number} filtered sinogram(s)

Sinograms.fbp2_back — Method

img = fbp2_back(sg, ig, sino; ia_skip)

2D backprojection for FBP.

in

sg::SinoGeom

ig::ImageGeom

sino::AbstractArray{<:Number} sinogram(s) (line integrals), usually ramp filtered

options

ia_skip::Int downsample in angle to save time for quick tests (default: 1)

out

img::AbstractArray{<:Number} reconstructed image(s)

Sinograms.fbp2_back_fan — Method

img = fbp2_back_fan(sg, ig, sino; ia_skip)

2D backprojection for fan-beam FBP.

in

sg::SinoGeom

ig::ImageGeom

sino::AbstractArray{<:Number} sinogram(s) (line integrals)

options

ia_skip::Int downsample in angle to save time for quick tests (default: 1)

out

img::AbstractArray{<:Number} reconstructed image(s)

Sinograms.fbp2_sino_filter — Method

sino, Hk, hh, nn = fbp2_sino_filter(how, sino;
ds=1, dsd=0, extra=0, npad=0, decon1=1, window=:none)

Apply ramp-like filters to sinogram(s) for 2D FBP image reconstruction. Both parallel-beam and fan-beam tomographic geometries are supported. This approach of sampling the band-limited ramp avoids the aliasing that would be caused by sampling the ramp directly in the frequency domain.

in

how::Symbol :arc (3rd generation CT) or :flat (for parallel too)

sino::AbstractArray{<:Number} : [nb (L)] sinograms

options

ds::RealU : sample spacing (in distance units, e.g., cm) (default: 1)

dsd::RealU : source-to-detector distance, for :arc case only. (default: Inf)

extra::Int : # of extra sinogram radial samples to keep (default: 0)

npad::Int : # of padded samples. (default: 0, means next power of 2)

decon1::Int : deconvolve effect of linear interpolator? (default: 1)

window::Symbol : default: :none

out

sino::AbstractArray{<:Number} filtered sinogram rows

Hk::AbstractVector{<:Number} apodized ramp filter frequency response

hn::AbstractVector{<:Number} samples of band-limited ramp filter

nn::AbstractVector{<:Number} [-np/2,...,np/2-1] vector for convenience

Sinograms.fbp2_sino_weight — Method

sino = fbp2_sino_weight(sg, ig, sino)

Apply sinogram weighting for first step of 2D fan-beam FBP.

in

sg::SinoGeom

sino::AbstractMatrix{<:Number}

out

sino::AbstractMatrix{<:Number}

Sinograms.fbp2_window — Method

window = fbp2_window(n, window)

Create an apodizing window of length n and fftshift it

Sinograms.fbp_ramp — Method

h, nn = fbp_ramp(how, n, ds, dsd)

'ramp-like' filters for parallel-beam and fan-beam FBP reconstruction. This sampled band-limited approach avoids the aliasing that would be caused by sampling the ramp directly in the frequency domain.

in

how::Symbol :arc (3rd generation CT) or :flat (for parallel too)

n::Int : # of samples (must be even)

ds::RealU : sample spacing (in distance units, e.g., cm)

dsd::RealU : source-to-detector distance, for :arc case

out

h::AbstractMatrix{<:Number} : samples of band-limited ramp filter