Methods list

Sinograms.Sinograms

Sinograms.CtFanArc

Sinograms.CtFanArc

Sinograms.CtFanArc

Sinograms.CtFanFlat

Sinograms.CtFanFlat

Sinograms.CtGeom

Sinograms.CtPar

Sinograms.CtPar

Sinograms.CtSource

Sinograms.CtSourceCircle

Sinograms.CtSourceHelix

Sinograms.CtSourceHelix

Sinograms.FBPNormalPlan

Sinograms.FBPplan

Sinograms.FDKplan

Sinograms.RayGeom

Sinograms.SinoFanArc

Sinograms.SinoFanArc

Sinograms.SinoFanArc

Sinograms.SinoFanFlat

Sinograms.SinoFanFlat

Sinograms.SinoGeom

Sinograms.SinoMoj

Sinograms.SinoMoj

Sinograms.SinoPar

Sinograms.SinoPar

Sinograms.Window

Sinograms.WindowVect

Base.show

Sinograms._geom_gamma

Sinograms._geom_rfov

Sinograms._geom_unitv

Sinograms._geom_xds

Sinograms._geom_yds

Sinograms._reale

Sinograms._show

Sinograms.cb_arc_to_par

Sinograms.cb_flat_to_par

Sinograms.cbct_back

Sinograms.cbct_back_fan!

Sinograms.ct_geom_plot2!

Sinograms.ct_geom_plot3

Sinograms.downsample

Sinograms.downsample

Sinograms.fbp

Sinograms.fbp

Sinograms.fbp!

Sinograms.fbp_back

Sinograms.fbp_back_fan

Sinograms.fbp_back_fan!

Sinograms.fbp_back_fan_xy

Sinograms.fbp_back_par

Sinograms.fbp_back_par!

Sinograms.fbp_back_par_xy

Sinograms.fbp_filter

Sinograms.fbp_ramp

Sinograms.fbp_sino_filter

Sinograms.fbp_sino_weight

Sinograms.fbp_window

Sinograms.fdk

Sinograms.footprint_size

Sinograms.oversample

Sinograms.oversample

Sinograms.parker_weight

Sinograms.plan_fbp

Sinograms.plan_fbp

Sinograms.rays

Sinograms.rays

Sinograms.reshaper

Sinograms.sino_geom_plot!

Sinograms.sino_geom_plot!

Sinograms.sino_geom_tau

Sinograms.sino_plot_rays

Sinograms.to_radians

Sinograms.to_radians

Sinograms.zwart_powell

Methods usage

Sinograms.Sinograms — Module

Sinograms module

Sinograms.CtFanArc — Type

CtFanArc 3D CBCT geometry for arc detector

Sinograms.CtFanArc — Method

CtFanArc(::Val{:ge1} ; kwargs...)

GE Lightspeed system CT geometry.

option

unit::RealU = 1 or use 1mm

out

CtFanArc

These numbers are published in IEEE T-MI Oct. 2006, p.1272-1283 wang:06:pwl.

julia> CtFanArc(Val(:ge1))
CtFanArc{Float64, Float32, CtSourceCircle} :
 ns::Int64 888
 nt::Int64 64
 ds::Float64 1.0239
 dt::Float64 1.0964
 offset_s::Float32 1.25
 offset_t::Float32 0.0
 na::Int64 984
 orbit::Float32 360.0
 orbit_start::Float32 0.0
 source_offset::Float64 0.0
 dsd::Float64 949.075
 dod::Float64 408.075
 src::CtSourceCircle CtSourceCircle()

Sinograms.CtFanArc — Method

CtFanArc( ; ns nt ds dt offset_s offset_t
    na orbit orbit_start
    dsd = 4ns * ds, dod = ns * ds)

Constructor with named keywords. See ?CtGeom for documentation.

julia> CtFanArc()
CtFanArc{Float32, Float32, CtSourceCircle} :
 ns::Int64 128
 nt::Int64 64
 ds::Float32 1.0
 dt::Float32 1.0
 offset_s::Float32 0.0
 offset_t::Float32 0.0
 na::Int64 64
 orbit::Float32 360.0
 orbit_start::Float32 0.0
 source_offset::Float32 0.0
 dsd::Float32 512.0
 dod::Float32 128.0
 src::CtSourceCircle CtSourceCircle()

Sinograms.CtFanFlat — Type

CtFanFlat 3D CTCT geometry for flat detector

Sinograms.CtFanFlat — Method

CtFanFlat( ; ns nt ds dt offset_s offset_t
    na orbit orbit_start
    dsd = 4ns * ds, dod = ns * ds)

Constructor with named keywords. See ?CtGeom for documentation.

julia> CtFanFlat()
CtFanFlat{Float32, Float32, CtSourceCircle} :
 ns::Int64 128
 nt::Int64 64
 ds::Float32 1.0
 dt::Float32 1.0
 offset_s::Float32 0.0
 offset_t::Float32 0.0
 na::Int64 64
 orbit::Float32 360.0
 orbit_start::Float32 0.0
 source_offset::Float32 0.0
 dsd::Float32 512.0
 dod::Float32 128.0
 src::CtSourceCircle CtSourceCircle()

Sinograms.CtGeom — Type

CtGeom{Td,To,Ts}

Abstract type for representing ray geometries for 3D CT imaging.

The projection view coordinates are (s,t) where s denotes the transaxial sampling and t denotes the axial direction (along z).

Common fields

ns size of each projection view

nt

ds detector pixel spacing

dt

offset_s unitless detector center offset (usually 0 or 0.25)

offset_t unitless, usually 0

na # of projection views, aka nϕ or nβ

orbit source orbit in degrees (or Unitful)

orbit_start starting angle in degrees (or Unitful) orbit and orbit_start must both be unitless (degrees) or have same units.

src::CtSource describes the X-ray CT source trajectory. Primary support for CtSourceCircle().

Additional fields for CtFan types:

dsd distance from source to detector

dod distance from origin to detector

source_offset usually 0

Units:

ds, dt, source_offset, dsd, dod must all be unitless or have the same units.

Derived values (available by getproperty), i.e., st.?

.dim dimensions: (ns,nt,na)

.s (ns) s sample locations

.t (nt) t sample locations

.ws = (ns-1)/2 + offset_s "middle" sample position

.wt = (nt-1)/2 + offset_t

.ad (na) source angles [degrees]

.ar (na) source angles [radians]

.rad (ns) radial distance of each ray from origin

.rfov max radius within FOV

.zfov axial FOV

For fan beam:

dso = dsd - dod distance from source to origin (Inf for parallel beam)

dfs distance from source to detector focal spot (0 for 3rd gen CT, Inf for flat detectors)

.gamma (ns) gamma sample values radians

.gamma_s gamma values given s values

.gamma_max half of fan angle radians, if offset_s=0

.gamma_max_abs half of fan angle radians - recommended

.cone_angle (half) cone angle on axis (s=0): +/- angle

Basic methods

dims (ns, nt, na)

ones = ones(Float32, ns,nt,na)

zeros = zeros(Float32, ns,nt,na)

rays iterator of (u,v,ϕ,θ) samples

downsample(st, down) reduce sampling by integer factor

oversample(st, over)

ct_geom_plot! plot system geometry

Methods

.shape(proj) reshape proj

.plot2(ig) plot the 2D geometry of the image

.plot3(ig) plot the 3D geometry of the cone beam setup

.unitv(is,it,ia) unit vector with single nonzero element

Notes

Use sino_geom() instead for 2D geometries.

Sinograms.CtPar — Type

CtPar

3D parallel-beam projection geometry

Sinograms.CtPar — Method

CtPar( ; ns nt ds dt offset_s offset_t na orbit orbit_start)

Constructor with named keywords. See ?CtGeom for documentation.

julia> CtPar()
CtPar{Float32, Float32, CtSourceCircle} :
 ns::Int64 128
 nt::Int64 64
 ds::Float32 1.0
 dt::Float32 1.0
 offset_s::Float32 0.0
 offset_t::Float32 0.0
 na::Int64 60
 orbit::Float32 180.0
 orbit_start::Float32 0.0
 src::CtSourceCircle CtSourceCircle()

Sinograms.CtSource — Type

CtSource

Abstract type for representing X-ray CT source trajectories.

Sinograms.CtSourceCircle — Type

CtSourceCircle

Type for representing circular X-ray CT source trajectory.

Sinograms.CtSourceHelix — Type

CtSourceHelix

Type for representing helical X-ray CT source trajectories having constant pitch.

Fields

pitch helix pitch (unitless fraction of zFOV, default 0)

source_z0 initial z position of x-ray source (default 0) It should have same units as detector pixels etc.

Sinograms.CtSourceHelix — Method

CtSourceHelix( ; pitch = 0, source_z0 = 0)

Constructor with named keywords

Sinograms.FBPNormalPlan — Type

FBPNormalPlan{S,I,H,P}

Struct type for storing "normal" FBP plan.

Sinograms.FBPplan — Type

FBPplan

Abstract type for FBP plans.

Sinograms.FDKplan — Type

FDKplan{C,I,H,P}

Struct type for storing FDK plan.

Sinograms.RayGeom — Type

RayGeom{Td,To}

Parent type for SinoGeom and CtGeom

Sinograms.SinoFanArc — Type

SinoFanArc 2D fan-beam sinogram geometry for arc detector

Sinograms.SinoFanArc — Method

SinoFan(Val(:ge1) ; kwargs...)

GE Lightspeed system CT geometry.

option

unit::RealU = 1 or use 1mm

nb::Int = 888 # of detector channels

d::RealU = 1.0239 channel spacing

offset::Real = 1.25 for "quarter-detector" offset

na::Int = 984 # of angles

orbit::Union{Symbol,Real} = 360 use :short for short scan

dsd::RealU = 949.075

dod::RealU = 408.075

out

SinoFanArc

These numbers are published in IEEE T-MI Oct. 2006, p.1272-1283 wang:06:pwl.

julia> SinoFanArc(Val(:ge1))
SinoFanArc{Float64, Float32} :
 nb::Int64 888
 d::Float64 1.0239
 offset::Float32 1.25
 na::Int64 984
 orbit::Float32 360.0
 orbit_start::Float32 0.0
 source_offset::Float64 0.0
 dsd::Float64 949.075
 dod::Float64 408.075

Sinograms.SinoFanArc — Method

SinoFanArc( ; nb d offset na orbit orbit_start
    source_offset, dsd = 4 * nb * d, dod = nb * d)
SinoFanArc(:short ; nb d offset na orbit_start
    source_offset, dsd = 4 * nb * d, dod = nb * d)

Constructor with named keywords. See ?SinoGeom for documentation.

d, source_offset, dsd, dod must all have the same units.

Use :short argument to specify a short scan, in which case na will be scaled down proportionally as well.

julia> SinoFanArc()
SinoFanArc{Float32, Float32} :
 nb::Int64 128
 d::Float32 1.0
 offset::Float32 0.0
 na::Int64 200
 orbit::Float32 360.0
 orbit_start::Float32 0.0
 source_offset::Float32 0.0
 dsd::Float32 512.0
 dod::Float32 128.0

Sinograms.SinoFanFlat — Type

SinoFanFlat

2D fan-beam sinogram geometry for flat detector

Sinograms.SinoFanFlat — Method

SinoFanFlat( ; nb d offset na orbit orbit_start
    source_offset, dsd= 4 * nb * d, dod = nb * d)

Constructor with named keywords. See ?SinoGeom for documentation.

julia> SinoFanFlat()
SinoFanFlat{Float32, Float32} :
 nb::Int64 128
 d::Float32 1.0
 offset::Float32 0.0
 na::Int64 200
 orbit::Float32 360.0
 orbit_start::Float32 0.0
 source_offset::Float32 0.0
 dsd::Float32 512.0
 dod::Float32 128.0

Sinograms.SinoGeom — Type

SinoGeom{Td,To}

Abstract type for representing ray geometries for 2D sinograms. This describes the sampling characteristics of a given sinogram for a 2D parallel or fan-beam system.

Common fields

nb # of "radial" samples, aka nr or ns

d aka dr or ds, "radial" sample spacing

offset unitless sample offset (usually 0 or 0.25)

na # of angular samples, aka nϕ or nβ default: 2 * floor(Int, nb * π/2 / 2)

orbit source orbit in degrees (or Unitful)

orbit_start starting angle in degrees (or Unitful) orbit and orbit_start must both be unitless (degrees) or have same units.

Additional fields for SinoFan types

dsd distance from source to detector

dod distance from origin to detector

source_offset usually 0

Units:

d, source_offset, dsd, dod must all have the same units.

Derived values (available by getproperty), i.e., st.?

.ds|dr radial sample spacing (NaN for :moj)

.s (nb) s sample locations

.w = (nb-1)/2 + offset "middle" sample position

.ad (na) source angles degrees

.ar (na) source angles radians

.rfov radial FOV

.xds (nb) center of detector elements (beta=0)

.yds (nb) ""

For mojette:

.d_ang (na) angle-dependent radial spacing

For fan beam:

dso = dsd - dod distance from source to origin (Inf for parallel beam)

dfs distance from source to detector focal spot (0 for 3rd gen CT, Inf for flat detectors)

.gamma (nb) gamma sample values radians

.gamma_max half of fan angle radians, if offset=0

Basic methods

dims (nb, na)

sino_s (nb) s sample locations

sino_w = (nb-1)/2 + offset ('middle' sample position)

rays iterator of (r, ϕ) parallel-beam coordinate tuples of size (nb, na)

ones = ones(Float32, nb, na)

zeros = zeros(Float32, nb, na)

downsample(st, down)

oversample(st, over)

angles (na) in degrees

sino_geom_plot! plot system geometry

Methods

.shape(sino) reshape sinograms into array (nb,na,:)

.unitv(;ib,ia) unit 'vector' with single nonzero element

.taufun(x,y) projected s/ds for each (x,y) pair (numel(x),na)

Notes

Use ct_geom() instead for 3D axial or helical cone-beam CT.

Sinograms.SinoMoj — Type

SinoMoj

2D Mojette sinogram geometry where d means dx (square pixel size)

Sinograms.SinoMoj — Method

SinoMoj( ; nb d offset na orbit orbit_start)

Constructor with named keywords. See ?SinoGeom for documentation.

julia> SinoMoj()
SinoMoj{Float32, Float32} :
 nb::Int64 128
 d::Float32 1.0
 offset::Float32 0.0
 na::Int64 200
 orbit::Float32 180.0
 orbit_start::Float32 0.0

Sinograms.SinoPar — Type

SinoPar

2D parallel-beam sinogram geometry

Sinograms.SinoPar — Method

SinoPar( ; nb d offset na orbit orbit_start)

Constructor with named keywords. See ?SinoGeom for documentation.

julia> SinoPar()
SinoPar{Float32, Float32} :
 nb::Int64 128
 d::Float32 1.0
 offset::Float32 0.0
 na::Int64 200
 orbit::Float32 180.0
 orbit_start::Float32 0.0

Sinograms.Window — Type

Window{S,T}

Data type for FBP apodizing windows, with given window shape::S and cutoff::T.

julia> Window(Hamming(), 0.8)
Window{Hamming, Float64}(Hamming(), 0.8)

Sinograms.WindowVect — Type

WindowVect{V} <: AbstractWindowShape

A user-specified window vector, constructed via WindowVect(v::V), where v is a AbstractVector. Caution: length(v) must be appropriate for the padded sinogram size.

Base.show — Method

show(io::IO, ::MIME"text/plain", st::RayGeom)

Sinograms._geom_gamma — Method

_geom_gamma(st::RayGeom)

Return gamma (γ: fan angle) values for fan-beam geometry.

Sinograms._geom_rfov — Method

_geom_rfov(st::RayGeom)

Radial FOV.

Sinograms._geom_unitv — Function

_geom_unitv([T=Float32], st:RayGeom [, pos::Tuple])

Projection views with a single non-zero ray value at position pos (default: middle).

Sinograms._geom_xds — Method

_geom_xds(st::RayGeom)

Center x positions of detectors (for beta = 0), for central row of detector.

Sinograms._geom_yds — Method

_geom_yds(st::RayGeom)

Center y positions of detectors (for beta = 0), for central row of detector.

Sinograms._reale — Method

_reale(x ; rtol = …)

Return the real part of x, but warn if the imaginary part is too large. By default this uses a 100× larger rtol than isapprox due to FFT errors.

Sinograms._show — Method

_show(io::IO, ::MIME"text/plain", st::Any)

Informative way to show fields of a struct (composite type).

Sinograms.cb_arc_to_par — Method

u, v, azim, polar = cb_arc_to_par(s, t, β, dso, dod)

Convert from cone-beam arc (3rd gen CT) coordinates to parallel-beam coordinates.

in

s,t detector coordinates

β X-ray source angle, measured counter-clockwise from y axis

dso,dsd distances for the geometry

out

u,v transaxial and axial parallel-beam detector coordinates

azim ϕ transaxial or azimuthal angle (radians)

polar θ polar angle (radians)

Sinograms.cb_flat_to_par — Method

u, v, azim, polar = cb_flat_to_par(s, t, β, dso, dod)

Convert from cone-beam flat panel coordinages to parallel-beam coordinates.

in

s,t detector coordinates

β X-ray source angle, measured counter-clockwise from y axis

dso,dsd distances for the geometry

out

u,v transaxial and axial parallel-beam detector coordinates

azim ϕ transaxial or azimuthal angle (radians)

polar θ polar angle (radians)

Sinograms.cbct_back — Method

cbct_back(proj, cg, ig)

Cone-beam backprojector for feldkamp.jl

in

proj (ns,nt,na) cone-beam projection views

cg::CtGeom

ig::ImageGeom

out

img (nx,ny,nz) back-projection result

Sinograms.cbct_back_fan! — Method

cbct_back_fan!(...)

This should work even for non-uniformly spaced source angles (for a circular source trajectory).

Sinograms.ct_geom_plot2! — Method

ct_geom_plot2!(st::CtGeom [, ig::ImageGeom])

Plot central 2D portion of the CBCT geometry.

Sinograms.ct_geom_plot3 — Method

ct_geom_plot3(st::CtFan [, ig::ImageGeom])

3D CBCT geometry plot

Sinograms.downsample — Method

downsample(ct, down::Int)
downsample(ct, down::NTuple{3,Int})

Down-sample CT geometry (for testing with small problems).

Sinograms.downsample — Method

downsample(st, down)

Down-sample sinogram geometry (for testing with small problems).

Sinograms.fbp! — Method

fbp!(image, sino ; orbit::Real = 180, kwargs...)

FBP reconstruction from a parallel-beam sinogram of size [nr × nϕ]. Writes result into image matrix.

Input

sino::AbstractMatrix

Options for SinoPar constructor

dr : sinogram radial spacing; default 1

orbit : angular range in degrees; default 180

orbit_start : angular range in degrees; default 0

Options for ImageGeom

dx, dy, deltas, offset_x, offset_y, offsets

rmax maximum radius for mask

Options

kwargs : passed to plan_fbp

Output

image::AbstractMatrix is mutated

Sinograms.fbp — Method

image, sino_filt = fbp(plan, sino)

Filtered back-projection (FBP) reconstruction, returning image and filtered sinogram.

in

plan::FBPplan

sino::AbstractMatrix{<:Number}

out

image::Matrix{<:Number} reconstructed image(s)

sino_filt::Matrix{<:Number} filtered sinogram(s)

Sinograms.fbp — Method

fbp(sino::AbstractMatrix ; kwargs...)

FBP reconstruction from a parallel-beam sinogram of size [nr × nϕ]. Returns an image of size [nx × ny].

Options

nx : default nr

ny : default nx

kwargs : passed to fbp!

Sinograms.fbp_back — Function

img = fbp_back(sg, ig, sino ; ia_skip)

2D pixel-driven backprojection for FBP.

in

sg::SinoGeom

ig::ImageGeom

sino::AbstractArray{<:Number} sinogram(s) (line integrals), usually ramp filtered

options

ia_skip::Int downsample in angle to save time for quick tests (default: 1)

out

img::Array{<:Number} reconstructed image(s)

Sinograms.fbp_back_fan! — Method

fbp_back_fan!(image, sino, sinβ, cosβ,
    dsd, dso, source_offset, is_arc,
    ds, offset, xc, yc, mask ; ia_skip)

Mutating version of pixel-driven back-projection for a grid of (xc,yc) pixel center locations for sinogram sino from a fan-beam geometry. It uses Threads. It assumes the angles are equally spaced over [0,π).

in

sino::Matrix{<:Number} (nb, na) usually ramp-filtered

sinβ::Vector{<:Real} (na)

cosβ::Vector{<:Real} (na)

dsd,dso,source_offset::RealU geometry

is_arc::Bool

ds::RealU ray spacing

offset::Real detector offset (usually 0)

xc::Vector{<:RealU} (nx) pixel centers

yc::Vector{<:RealU} (ny) pixel centers

mask::Matrix{Bool} (nx, ny) which pixels to reconstruct

option

ia_skip::Int default 1

out

image::Matrix (nx, ny) matrix to be mutated

Sinograms.fbp_back_fan — Method

fbp_back_fan(sino, betas,
    dsd, dso, dfs, source_offset, is_arc,
    ds, offset, xc, yc, mask ; ia_skip, T)

Pixel-driven back-projection for a grid of (xc,yc) pixel center locations for sinogram sino from a fan-beam geometry. It assumes the angles are equally spaced over [0,π).

in

sino::Matrix{<:Number} (nb, na) usually ramp-filtered

betas::Vector{<:Real} (na) in radians

dsd,dso,source_offset::RealU geometry

is_arc::Bool arc or flat?

ds::RealU ray spacing

offset::Real detector offset (usually 0)

xc::Vector{<:RealU} (nx) pixel centers

yc::Vector{<:RealU} (ny) pixel centers

mask::Matrix{Bool} (nx, ny) which pixels to reconstruct

option

ia_skip::Int default 1

T::Type{<:Number} usually same as eltype(sino)

out

image::Matrix (nx, ny)

Sinograms.fbp_back_fan_xy — Method

fbp_back_fan_xy(sino, sinβ, cosβ,
    dsd_ds, dso_ds, source_offset_ds, is_arc,
    wb, x, y ; T)

Pixel-driven back-projection for a single (x,y) location for sinogram sino from a fan-beam geometry. It assumes the angles are equally spaced over [0,π).

in

sino::Matrix{<:Number} (nb, na) usually ramp-filtered

sinβ::Vector{<:Real} (na)

cosβ::Vector{<:Real} (na)

dsd_ds,dso_ds,source_offset_ds::Real geometry, normalized

wb::Real = (nb+1)/2 + offset where usually offset=0

x,y::Real pixel center location, normalized by ray spacing

option

T::Type{<:Number} typically same as eltype(sino)

out

Returns a scalar of type T.

Sinograms.fbp_back_par! — Method

fbp_back_par!(image, sino, sinϕ, cosϕ,
    ds, offset, xc, yc, mask ; ia_skip)

Mutating version of pixel-driven back-projection for a grid of (xc,yc) pixel center locations for sinogram sino from a parallel-beam geometry. It uses Threads. It assumes the angles are equally spaced over [0,π).

in

sino::Matrix{<:Number} (nb, na) usually ramp-filtered

sinϕ::Vector{<:Real} (na)

cosϕ::Vector{<:Real} (na)

ds::RealU ray spacing

offset::Real detector offset (usually 0)

xc::Vector{<:RealU} (nx) pixel centers

yc::Vector{<:RealU} (ny) pixel centers

mask::Matrix{Bool} (nx, ny) which pixels to reconstruct

option

ia_skip::Int default 1

out

image::Matrix (nx, ny) matrix to be mutated

Sinograms.fbp_back_par — Method

fbp_back_par(sino, angles,
    ds, offset, xc, yc, mask ; ia_skip, T)

Pixel-driven back-projection for a grid of (xc,yc) pixel center locations for sinogram sino from a parallel-beam geometry. It assumes the angles are equally spaced over [0,π).

in

sino::Matrix{<:Number} (nb, na) usually ramp-filtered

angles::Vector{<:Real} (na) in radians

ds::RealU ray spacing

offset::Real detector offset (usually 0)

xc::Vector{<:RealU} (nx) pixel centers

yc::Vector{<:RealU} (ny) pixel centers

mask::Matrix{Bool} (nx, ny) which pixels to reconstruct

option

ia_skip::Int default 1

T::Type{<:Number} usually same as eltype(sino)

out

image::Matrix (nx, ny)

Sinograms.fbp_back_par_xy — Method

fbp_back_par_xy(sino, sinϕ, cosϕ,
    wb, x, y ; T)

Pixel-driven back-projection for a single (x,y) location for sinogram sino from a parallel-beam geometry. It assumes the angles are equally spaced over [0,π).

in

sino::Matrix{<:Number} (nb, na) usually ramp-filtered

sinϕ::Vector{<:Real} (na)

cosϕ::Vector{<:Real} (na)

wb::Real = (nb+1)/2 + offset where usually offset=0

x,y::Real pixel center location, normalized by ray spacing

option

T::Type{<:Number} typically same as eltype(sino)

out

Returns a scalar of type T.

Sinograms.fbp_filter — Method

Hk = fbp_filter(st::Union{SinoGeom,CtGeom} ;
    npad=0, ds::RealU = st.d, decon1::Bool=true, window=Window())

Compute frequency response of ramp-like filter used for FBP image reconstruction. Supports parallel-beam and fan-beam tomographic geometries in 2D and 3D. This code samples the band-limited ramp to avoid the aliasing that would be caused by sampling the ramp directly in the frequency domain.

in

st::Union{SinoGeom,CtGeom}

options

npad::Int # of padded samples. (default: next power of 2)

ds::Td detector sample spacing (default from st)

decon1::Bool deconvolve effect of linear interpolator? (default: true)

window::Window apodizer; default: Window()

out

Hk::Vector apodized ramp filter frequency response

Sinograms.fbp_ramp — Method

h, n = fbp_ramp(sg::SinoGeom, N::Int)

'ramp-like' filters for parallel-beam and fan-beam FBP reconstruction. This sampled band-limited approach avoids the aliasing that would be caused by sampling the ramp directly in the frequency domain.

in

sg::SinoGeom

N::Int : # of samples (must be even)

out

h::Vector{<:RealU} : samples of band-limited ramp filter

n::UnitRange{Int64} : -(N÷2):(N÷2-1)

Sinograms.fbp_sino_filter — Method

sino = fbp_sino_filter(sino::Array, filter::Vector ; extra=0)

Apply ramp-like filters to sinogram(s) for 2D FBP image reconstruction. Supports both parallel-beam and fan-beam tomographic geometries in 2D and 3D.

in

sino::AbstractArray{<:Number} [nb (L)] sinograms

filter::AbstractVector (npad ≥ nb) apodized ramp filter frequency response

options

extra::Int # of extra sinogram radial samples to keep (default: 0)

npad::Int # of padded samples. (default: next power of 2)

out

sino::AbstractArray sinogram with filtered rows

Sinograms.fbp_sino_weight — Method

fbp_sino_weight(sg::SinoFan)

Return 1D sinogram weighting for first step of 2D fan-beam FBP.

Sinograms.fbp_window — Method

fbp_window(w::Window, N::Int ; T = Float32)

Create an apodizing window of length N and fftshift it.

julia> fbp_window(Window(Hamming()), 4)
4-element Vector{Float32}:
 1.0
 0.54
 0.0
 0.54

Sinograms.fdk — Method

image = fdk(plan, proj)

Reconstruct 3D image from cone-beam computed tomography (CBCT) data collected with a circular source trajectory via FDK method.

in

plan::FDKplan

proj (ns,nt,na) projection views

out

image (nx,ny,nz) reconstructed image

References: Feldkamp, Davis, Kress, JOSA-A, 1(6):612-9, June 1984.

Sinograms.footprint_size — Function

footprint_size(st::Union{SinoGeom,CtGeom}, ig::ImageGeom)

Unitless maximum footprint size (in detector pixels).

Sinograms.oversample — Method

oversample(st, over::Int)

Over-sample CT geometry in "radial" dimension.

Sinograms.oversample — Method

oversample(st, over::Int)

Over-sample sinogram geometry in "radial" dimension. For Mojette sampling, it means that d = dx/over.

Sinograms.parker_weight — Method

parker_weight(sg::SinoGeom ; T = Float32)

Compute Parker weighting for non-360° orbits. Returns Vector{T} of length sg.na for SinoPar. Returns Matrix{T} of size dims(sg) for SinoFan.

Sinograms.plan_fbp — Method

plan = plan_fbp(cg, ig; window=Window(), ...)

Plan FDK 3D CBCT image reconstruction, with either flat or arc detector.

To use this, you first call it with the CT geometry and image geometry. The routine returns the initialized plan. Thereafter, to to perform FDK reconstruction, call fbp with the plan (perhaps numerous times for the same geometry).

in

cg::CtGeom

ig::ImageGeom only reconstruct pixels within ig.mask.

options

window::Window e.g., Window(Hamming(), 0.8); default Window()

npad::Int # of radial bins after padding; default nextpow(2, cg.ns + 1)

decon1::Bool deconvolve interpolator effect? (default true)

-T::Type{<:Number} type of sino elements (default Float32)

out

plan::FDKplan initialized plan

Sinograms.plan_fbp — Method

plan = plan_fbp(sg, ig; how=:normal, window=Window())

Plan FBP 2D tomographic image reconstruction for parallel-beam & fan-beam cases, with either flat or arc detector for fan-beam case.

To use this, you first call it with the sinogram and image geometry. The routine returns the initialized plan. Thereafter, to to perform FBP reconstruction, call fbp with the plan (perhaps numerous times for the same geometry).

in

sg::SinoGeom

ig::ImageGeom only reconstruct pixels within ig.mask.

options

how::Symbol how to reconstruct

:normal default

:mojette use mojette rebinning and Gtomo2_table

window::Window e.g., Window(Hamming(), 0.8); default Window()

npad::Int # of radial bins after padding; default nextpow(2, sg.nb + 1)

decon1::Bool deconvolve interpolator effect? (default true)

T::Type{<:Number} type of sino elements (default Float32)

out

plan::FBPplan initialized plan

Sinograms.rays — Method

i = rays(st::CtGeom)

Return parallel-beam coordinates of all rays for this CT geometry. Return type of i is a ProductIterator that makes tuples of the form (u, v, ϕ, θ). To make projections call p = [fun(c...) for c in i] where fun is radon(...).

Sinograms.rays — Method

 = rays(sg::SinoGeom)

Radial r and angular ϕ coordinates (in radians) of all sinogram elements for the given geometry. Return type of i is a ProductIterator that makes tuples of the form (r, ϕ). To make projections call p = [fun(c...) for c in i] where fun is radon(...).

Sinograms.reshaper — Method

reshaper(x::AbstractArray, dim:Dims)

Reshape x to size dim with : only if needed

Sinograms.sino_geom_plot! — Method

sino_geom_plot!(sg, ig)

Picture of the source position / detector geometry.

Sinograms.sino_geom_plot! — Method

sino_geom_plot!(sg)

Picture of the source position / detector geometry.

Sinograms.sino_geom_tau — Method

sino_geom_tau(st::Union{SinoGeom,CtGeom}, x, y)

Projected s/ds, useful for footprint center and support. Returns Matrix of size length(x) × st.na.

Sinograms.sino_plot_rays — Method

sino_plot_rays(sg::SinoGeom ; kwargs...)

Make a scatter plot of the (r, ϕ) sample locations for all rays. Requires Plots.

Sinograms.to_radians — Method

to_radians(angles::AbstractArray{Unitful.Quantity})

Convert Unitful quantity array to radians.

Sinograms.to_radians — Method

to_radians(angles::AbstractArray{<:AbstractFloat})

When Unitful package not loaded, assume angles are in degrees and convert to radians.

Sinograms.zwart_powell — Method

output = zwart_powell(r, ϕ)

Analytic 2D Radon transform value of Zwart-Powell box spline, for radial distance r (normalized by pixel size) and angle ϕ (in radians).