Methods list

Sinograms.Sinograms

Sinograms.SinoFan

Sinograms.SinoFanArc

Sinograms.SinoFanArc

Sinograms.SinoFanFlat

Sinograms.SinoFanFlat

Sinograms.SinoGeom

Sinograms.SinoMoj

Sinograms.SinoMoj

Sinograms.SinoPar

Sinograms.SinoPar

Sinograms.Window

Sinograms.WindowVect

Base.show

Sinograms._reale

Sinograms.downsample

Sinograms.fbp

Sinograms.fbp

Sinograms.fbp!

Sinograms.fbp_back

Sinograms.fbp_filter

Sinograms.fbp_ramp

Sinograms.fbp_sino_filter

Sinograms.fbp_sino_weight

Sinograms.fbp_window

Sinograms.oversample

Sinograms.parker_weight

Sinograms.plan_fbp

Sinograms.rays

Sinograms.reshaper

Sinograms.sino_geom_gamma

Sinograms.sino_geom_rfov

Sinograms.sino_geom_taufun

Sinograms.sino_geom_unitv

Sinograms.sino_geom_xds

Sinograms.sino_geom_yds

Sinograms.sino_plot_rays

Sinograms.zwart_powell

Methods usage

Sinograms.Sinograms — Module

Sinograms module

Sinograms.SinoFan — Method

SinoFan(Val(:ge1))

Sinogram geometry for GE lightspeed system. These numbers are published in IEEE T-MI Oct. 2006, p.1272-1283 wang:06:pwl.

Sinograms.SinoFanArc — Type

SinoFanArc : 2D fan-beam sinogram geometry for arc detector

julia> SinoFanArc()
SinoFanArc{Float32, Float32} :
 nb::Int64 128
 na::Int64 200
 d::Float32 1.0
 orbit::Float32 360.0
 orbit_start::Float32 0.0
 offset::Float32 0.0
 strip_width::Float32 1.0
 source_offset::Float32 0.0
 dsd::Float32 512.0
 dod::Float32 128.0
 dfs::Float32 0.0

Sinograms.SinoFanArc — Method

SinoFanArc( ; nb, na, d=1, orbit=360, orbit_start=0, offset=0, strip_width=d,
    source_offset = 0, dsd= 4 * nb * d, dod = nb * d, dfs = 0, down=0)

Constructor with named keywords.

orbit and orbit_start must both be unitless (degrees) or have same units.

d, strip_width, source_offset, dsd, dod, dfs must all be unitless or have the same units.

nb::Int = 128 # of radial bins

na::Int = 2 * floor(Int, nb * π/2 / 2) # of projection angles

Sinograms.SinoFanFlat — Type

SinoFanFlat : 2D fan-beam sinogram geometry for flat detector

julia> SinoFanFlat()
SinoFanFlat{Float32, Float32} :
 nb::Int64 128
 na::Int64 200
 d::Float32 1.0
 orbit::Float32 360.0
 orbit_start::Float32 0.0
 offset::Float32 0.0
 strip_width::Float32 1.0
 source_offset::Float32 0.0
 dsd::Float32 512.0
 dod::Float32 128.0
 dfs::Float32 Inf

Sinograms.SinoFanFlat — Method

SinoFanFlat( ; nb, na, d=1, orbit=360, orbit_start=0, offset=0, strip_width=d,
    source_offset = 0, dsd= 4 * nb * d, dod = nb * d, dfs = Inf, down=0)

Constructor with named keywords.

orbit and orbit_start must both be unitless (degrees) or have same units.

d, strip_width, source_offset, dsd, dod, dfs must all be unitless or have the same units.

nb::Int = 128 # of radial bins

na::Int = 2 * floor(Int, nb * π/2 / 2) # of projection angles

Sinograms.SinoGeom — Type

SinoGeom

Abstract type for representing 2D sinogram ray geometries.

Basic methods:

dims (nb, na)

sino_s [nb] s sample locations

sino_w (nb-1)/2 + offset ('middle' sample position)

rays (rgrid, ϕgrid) [nb na] parallel-beam coordinates

ones ones(Float32, nb, na)

zeros zeros(Float32, nb, na)

downsample

oversample

angles (na) in degrees

Derived values (available by getproperty), i.e., sg.?

.dim dimensions: (nb,na)

.ds|dr radial sample spacing (NaN for :moj)

.s [nb] s sample locations

.w (nb-1)/2 + offset ('middle' sample position)

.ad source angles [degrees]

.ar source angles [radians]

.ones ones(Float32, nb,na)

.zeros zeros(Float32, nb,na)

.rfov radial fov

.xds [nb] center of detector elements (beta=0)

.yds [nb] ''

.plot_grid(scatter) plot sg.grid using Plots.scatter

For mojette:

.d_ang [na]

For fan beam:

.gamma [nb] gamma sample values [radians]

.gamma_max half of fan angle [radians]

.dso dsd - dod, Inf for parallel beam

dsd dis_src_det distance from source to detector

dso dis_src_iso distance from source to isocenter

dod dis_iso_det distance from isocenter to detector

dfs dis_foc_src distance from source to detector focal spot (0 for 3rd gen CT, Inf for flat detectors)

Methods

.down(down) reduce sampling by integer factor

.shape(sino) reshape sinograms into array [nb na :]

.unitv(;ib,ia) unit 'vector' with single nonzero element

.taufun(x,y) projected s/ds for each (x,y) pair [numel(x) na]

.plot!(plot!;ig)plot system geometry (mostly for SinoFan)

Sinograms.SinoMoj — Type

SinoMoj : 2D Mojette sinogram geometry
where `d` means `dx` (square pixel size), and `strip_width` may be ignored.

julia> SinoMoj()
SinoMoj{Float32, Float32} :
 nb::Int64 128
 na::Int64 200
 d::Float32 1.0
 orbit::Float32 180.0
 orbit_start::Float32 0.0
 offset::Float32 0.0
 strip_width::Float32 1.0

Sinograms.SinoMoj — Method

SinoMoj( ; nb, na, d=1, orbit=180, orbit_start=0, offset=0, strip_width=d, down=0)

Constructor with named keywords.

orbit and orbit_start must both be unitless (degrees) or have same units.

d and strip_width must both be unitless or have the same units.

d means dx (square pixel size)

strip_width may be ignored (todo).

nb::Int = 128 # of radial bins

na::Int = 2 * floor(Int, nb * π/2 / 2) # of projection angles

Sinograms.SinoPar — Type

SinoPar : 2D parallel-beam sinogram geometry

julia> SinoPar()
SinoPar{Float32, Float32} :
 nb::Int64 128
 na::Int64 200
 d::Float32 1.0
 orbit::Float32 180.0
 orbit_start::Float32 0.0
 offset::Float32 0.0
 strip_width::Float32 1.0

Sinograms.SinoPar — Method

SinoPar( ; nb, na, d=1, orbit=180, orbit_start=0, offset=0, strip_width=d, down=0)

Constructor with named keywords.

orbit and orbit_start must both be unitless (degrees) or have same units.

d and strip_width must both be unitless or have the same units.

nb::Int = 128 # of radial bins

na::Int = 2 * floor(Int, nb * π/2 / 2) # of projection angles

Sinograms.Window — Type

Window{S,T}

Data type for FBP apodizing windows.

Sinograms.WindowVect — Type

WindowVect{V} <: AbstractWindowShape

A user-specified window vector, constructed via WindowVect(v::V), where v is a AbstractVector. Caution: length(v) must be appropriate for the padded sinogram size.

Base.show — Method

show(io::IO, ::MIME"text/plain", sg::SinoGeom)

Sinograms._reale — Method

_reale(x ; rtol = …)

Return the real part of x, but warn if the imaginary part is too large. By default this uses a 100× larger rtol than isapprox due to FFT errors.

Sinograms.downsample — Method

sg = downsample(sg, down)

Down-sample sinogram geometry (for testing with small problems).

Sinograms.fbp! — Method

fbp!(image, sino ; orbit::Real = 180, kwargs...)

FBP reconstruction from a parallel-beam sinogram of size [nr × nϕ]. Writes result into image matrix.

Input

sino::AbstractMatrix

Options for SinoPar constructor

dr : sinogram radial spacing; default 1

orbit : angular range in degrees; default 180

orbit_start : angular range in degrees; default 0

Options for ImageGeom

dx, dy, deltas, offset_x, offset_y, offsets

rmax maximum radius for mask

Options

kwargs : passed to plan_fbp

Output

image::AbstractMatrix is mutated

Sinograms.fbp — Method

image, sino_filt = fbp(plan, sino)

Filtered back-projection (FBP) reconstruction, returning image and filtered sinogram.

in

plan::FBPplan

sino::AbstractMatrix{<:Number}

out

image::Matrix{<:Number} reconstructed image(s)

sino_filt::Matrix{<:Number} filtered sinogram(s)

Sinograms.fbp — Method

fbp(sino::AbstractMatrix ; kwargs...)

FBP reconstruction from a parallel-beam sinogram of size [nr × nϕ]. Returns an image of size [nx × ny].

Options

nx : default nr

ny : default nx

kwargs : passed to fbp!

Sinograms.fbp_back — Function

img = fbp_back(sg, ig, sino ; ia_skip)

2D pixel-driven backprojection for FBP.

in

sg::SinoGeom

ig::ImageGeom

sino::AbstractArray{<:Number} sinogram(s) (line integrals), usually ramp filtered

options

ia_skip::Int downsample in angle to save time for quick tests (default: 1)

out

img::Array{<:Number} reconstructed image(s)

Sinograms.fbp_filter — Method

Hk = fbp_filter(sg::SinoGeom ;
    npad=0, ds::RealU = sg.d, decon1::Bool=true, window=Window())

Compute frequency response of ramp-like filter used for 2D FBP image reconstruction. Both parallel-beam and fan-beam tomographic geometries are supported. This code samples the band-limited ramp to avoid the aliasing that would be caused by sampling the ramp directly in the frequency domain.

in

sg::SinoGeom

options

npad::Int # of padded samples. (default: next power of 2)

decon1::Bool deconvolve effect of linear interpolator? (default: true)

window::Window apodizer; default: Window()

out

Hk::Vector apodized ramp filter frequency response

Sinograms.fbp_ramp — Method

h, n = fbp_ramp(sg::SinoGeom, N::Int)

'ramp-like' filters for parallel-beam and fan-beam FBP reconstruction. This sampled band-limited approach avoids the aliasing that would be caused by sampling the ramp directly in the frequency domain.

in

sg::SinoGeom

N::Int : # of samples (must be even)

out

h::Vector{<:RealU} : samples of band-limited ramp filter

n::UnitRange{Int64} : -(N÷2):(N÷2-1)

Sinograms.fbp_sino_filter — Method

sino = fbp_sino_filter(sg::SinoGeom, sino, filter ; extra=0)

Apply ramp-like filters to sinogram(s) for 2D FBP image reconstruction. Both parallel-beam and fan-beam tomographic geometries are supported. todo?

in

sg::SinoGeom

sino::AbstractArray{<:Number} [nb (L)] sinograms

filter::AbstractVector (npad ≥ nb) apodized ramp filter frequency response

options

extra::Int # of extra sinogram radial samples to keep (default: 0)

npad::Int # of padded samples. (default: next power of 2)

out

sino::AbstractArray sinogram with filtered rows

Sinograms.fbp_sino_weight — Method

fbp_sino_weight(sg::SinoFan)

Return 1D sinogram weighting for first step of 2D fan-beam FBP.

Sinograms.fbp_window — Method

fbp_window(w::Window, N::Int ; T = Float32)

Create an apodizing window of length N and fftshift it.

Sinograms.oversample — Method

sg = oversample(sg, over::Int)

Over-sample sinogram geometry in "radial" dimension. For Mojette sampling, it means that d = dx/over.

Sinograms.parker_weight — Method

parker_weight(sg::SinoGeom ; T = Float32)

Compute Parker weighting for non-360° orbits. Returns Vector{T} of length sg.na.

Sinograms.plan_fbp — Function

plan = plan_fbp(sg, ig; how=:normal, window=Window())

Plan FBP 2D tomographic image reconstruction for parallel-beam & fan-beam cases, with either flat or arc detector for fan-beam case.

To use this, you first call it with the sinogram and image geometry. The routine returns the initialized "plan". Thereafter, to to perform FBP reconstruction, call fbp with the plan (perhaps numerous times for the same geometry).

in

sg::SinoGeom

ig::ImageGeom # only reconstruct pixels within ig.mask.

options

how::Symbol how to reconstruct

:normal default

:mojette use mojette rebinning and Gtomo2_table

window::Window e.g. Window(Hamming(0.5)); default Window()

= npad::Int # of radial bins after padding; default nextpow(2, sg.nb + 1) -T::DataType type of sino elements (default: Float32)

out

plan::FBPplan initialized plan

Sinograms.rays — Method

(r, ϕ) = rays(sg::SinoGeom)

Radial r and angular ϕ coordinates (in radians) of all sinogram elements for the given geometry.

Sinograms.reshaper — Method

reshaper(x::AbstractArray, dim:Dims)

Reshape x to size dim with : only if needed

Sinograms.sino_geom_gamma — Method

sino_geom_gamma(sg::SinoFan)

Return gamma (fan angle) values for fan-beam geometry

Sinograms.sino_geom_rfov — Method

sino_geom_rfov(sg::SinoGeom)

Radial FOV.

Sinograms.sino_geom_taufun — Method

sino_geom_taufun(sg::SinoGeom, x, y)

Projected s/ds, useful for footprint center and support. Returns Matrix of size length(x) × sg.na.

Sinograms.sino_geom_unitv — Method

sino_geom_unitv([T=Float32], sg:SinoGeom ; ib::Int, ia::Int)

Sinogram with a single ray

Sinograms.sino_geom_xds — Method

sino_geom_xds(sg::SinoGeom)

Center x positions of detectors (for beta = 0).

Sinograms.sino_geom_yds — Method

sino_geom_yds(sg::SinoGeom)

Center y positions of detectors (for beta = 0).

Sinograms.sino_plot_rays — Method

sino_plot_rays(sg::SinoGeom ; kwargs...)

Make a scatter plot of the (r, ϕ) sample locations for all rays. Requires Plots.

Sinograms.zwart_powell — Method

output = zwart_powell(r, ϕ)

Analytic 2D Radon transform value of Zwart-Powell box spline, for radial distance r (normalized by pixel size) and angle ϕ (in radians).